requirements.txt
absl-py==0.11.0
aiohttp==3.7.4.post0
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.1.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
colorama==0.4.4
flatbuffers==1.12
fsspec==2021.7.0
future==0.18.2
gast==0.4.0
google-auth==1.34.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.39.0
idna==3.2
Markdown==3.3.3
multidict==5.1.0
nflows==0.14
numpy==1.21.1
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==21.0
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pyparsing==2.4.7
pytorch-lightning==1.4.0
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.16.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
termcolor==1.1.0
torch==1.9.0
torchmetrics==0.4.1
torchvision==0.10.0
tqdm==4.61.2
timm==0.4.12
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.12.1
yarl==1.6.3
